reModify the Project folder
