+++ src/main/java/module-info.java
@@ -15,11 +15,20 @@ module ui.gp {
     exports ui.gp.SceneController;
     exports ui.gp.SceneController.Customer;
     exports ui.gp.SceneController.Customer.Dependent;
-    exports ui.gp.Utils;
+    exports ui.gp.Database;
     exports ui.gp.Models.Users;
 
     opens ui.gp to javafx.fxml;
     opens ui.gp.SceneController to javafx.fxml;
     opens ui.gp.SceneController.Customer to javafx.fxml;
     opens ui.gp.SceneController.Customer.Dependent to javafx.fxml;
+    exports ui.gp.Tab;
+    exports ui.gp.SceneController.Policy;
+    opens ui.gp.SceneController.Policy to javafx.fxml;
+    exports ui.gp.SceneController.Function;
+    opens ui.gp.SceneController.Function to javafx.fxml;
+    exports ui.gp.SceneController.Manager;
+    opens ui.gp.SceneController.Manager to javafx.fxml;
+    exports ui.gp.SceneController.Controllers;
+    opens ui.gp.SceneController.Controllers to javafx.fxml;
 }
