--- src/main/java/ui/gp/SceneController/Customer/PolicyHolder/listOfClaimController.java
@@ -1,4 +1,0 @@
-package ui.gp.SceneController.Customer.PolicyHolder;
-
-public class listOfClaimController {
-}
