+++ src/main/java/ui/gp/SceneController/Manager/AdminHomeController.java
@@ -1,0 +1,4 @@
+package ui.gp.SceneController.Manager;
+
+public class AdminHomeController {
+}
