+++ src/main/java/ui/gp/SceneController/Policy/OwnerHomeController.java
@@ -1,0 +1,29 @@
+package ui.gp.SceneController.Policy;
+
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.layout.AnchorPane;
+import ui.gp.SceneController.Function.SceneUtil;
+
+import java.io.IOException;
+
+public class OwnerHomeController {
+    @FXML
+    Label welcomeBannerUser;
+    @FXML
+    AnchorPane ownerHomeScene;
+    @FXML
+    Button logoutButton;
+
+
+    public void bannerNameView(String username) {
+        welcomeBannerUser.setText("Welcome " + username);
+    }
+
+    @FXML
+    public void logoutOwner(ActionEvent logoutAction) throws IOException {
+        SceneUtil.logout(ownerHomeScene);
+    }
+}
