+++ src/main/java/ui/gp/Tab/listOfClaimController.java
@@ -1,0 +1,4 @@
+package ui.gp.Tab;
+
+public class listOfClaimController {
+}
