--- src/main/java/ui/gp/View/ViewFactory.java
@@ -1,4 +1,0 @@
-package ui.gp.View;
-
-public class ViewFactory {
-}
