--- src/main/resources/ui/gp/Scene/Customer/DependentsView/claim.fxml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<AnchorPane xmlns="http://javafx.com/javafx"
-            xmlns:fx="http://javafx.com/fxml"
-            fx:controller="ui.gp.Scene.Customer.DependentsView.Claim"
-            prefHeight="400.0" prefWidth="600.0">
-
-</AnchorPane>
