+++ src/main/resources/ui/gp/Scene/Function/LoadingScreen.fxml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Font?>
+
+<fx:root fx:id="loadingPane" prefHeight="307.0" prefWidth="492.0" type="AnchorPane" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ui.gp.SceneController.Function.LoadingSceneController">
+   <children>
+      <Label layoutX="206.0" layoutY="215.0" text="Loading..." AnchorPane.bottomAnchor="64.0" AnchorPane.leftAnchor="206.0" AnchorPane.rightAnchor="206.31068420410156" AnchorPane.topAnchor="215.0">
+         <font>
+            <Font size="19.0" />
+         </font>
+      </Label>
+      <ImageView fitHeight="150.0" fitWidth="200.0" layoutX="155.0" layoutY="31.0" pickOnBounds="true" preserveRatio="true" AnchorPane.bottomAnchor="126.0" AnchorPane.leftAnchor="155.0" AnchorPane.rightAnchor="155.68931579589844" AnchorPane.topAnchor="31.0">
+         <image>
+            <Image url="@../../Image/touhou-spinning.gif" />
+         </image>
+      </ImageView>
+   </children>
+</fx:root>
