+++ src/main/resources/ui/gp/Scene/Function/ProofUploadScene.fxml
